Adding specific class defining the ldp_container taxonomy
