+++ wpldp.php
@@ -13,6 +13,7 @@
 // If the file is accessed outside of index.php (ie. directly), we just deny the access
 defined('ABSPATH') or die("No script kiddies please!");
 
+require_once('wpldp-taxonomy.php');
 require_once('wpldp-settings.php');
 
 if (!class_exists('WpLdp')) {
@@ -35,31 +36,17 @@ if (!class_exists('WpLdp')) {
         add_action( 'init', array($this, 'register_connection_types'));
         add_action( 'edit_form_after_title', array($this, 'myprefix_edit_form_after_title'));
         add_action( 'save_post', array($this, 'test_save'));
-        //add_action('update_option', 'initialize_container');
 
         add_filter( 'template_include', array($this, 'include_template_function'));
         add_action( 'template_redirect', array($this, 'my_page_template_redirect' ));
         add_action( 'add_meta_boxes', array($this, 'display_container_meta_box' ));
-        add_action( 'init', array($this, 'register_container_taxonomy'), 0 );
 
-        add_action( 'ldp_container_add_form_fields', array($this, 'add_custom_tax_fields_oncreate'));
-        add_action( 'ldp_container_edit_form_fields', array($this, 'add_custom_tax_fields_onedit'));
-        add_action( 'create_ldp_container', array($this, 'save_custom_tax_field'));
-        add_action( 'edited_ldp_container', array($this, 'save_custom_tax_field'));
         add_action( 'admin_footer', array($this, 'my_action_javascript' )); // Write our JS below here
 
         add_filter( 'post_type_link', array($this, 'ldp_resource_post_link'), 10, 3 );
 
         add_action('admin_enqueue_scripts', array($this, 'ldp_enqueue_script'));
         add_action('admin_enqueue_scripts', array($this, 'ldp_enqueue_stylesheet'));
-      }
-
-
-      function initialize_container($option, $oldValue, $_newValue) {
-        if ($option === 'ldp_container_init') {
-          var_dump("This is the good option update");
-          die();
-        }
       }
 
       #####################################
@@ -243,16 +230,7 @@ if (!class_exists('WpLdp')) {
                   }
               }
           }
-          elseif ( is_tax('ldp_container') ) {
-              // checks if the file exists in the theme first,
-              // otherwise serve the file from the plugin
-              if ( $file_theme = locate_template( array ( 'taxonomy-ldp_resource.php' ) ) ) {
-                  $template_path = $theme_file;
-              } else {
-                  $template_path = plugin_dir_path( __FILE__ ) . 'taxonomy-ldp_resource.php';
-              }
 
-          }
           return $template_path;
       }
 
@@ -342,128 +320,6 @@ if (!class_exists('WpLdp')) {
             plugins_url('library/js/node_modules/jsoneditor/dist/jsoneditor.min.css', __FILE__)
           );
           wp_enqueue_style('jsoneditorcss');
-      }
-
-      ################################
-      # Taxonomy
-      ################################
-      // Register Custom Taxonomy
-      function register_container_taxonomy() {
-
-      	$labels = array(
-      		'name'                       => __( 'Containers', 'wpldp' ),
-      		'singular_name'              => __( 'Container', 'wpldp' ),
-      		'menu_name'                  => __( 'Containers', 'wpldp' ),
-      		'all_items'                  => __( 'All Items', 'wpldp' ),
-      		'parent_item'                => __( 'Parent Item', 'wpldp' ),
-      		'parent_item_colon'          => __( 'Parent Item:', 'wpldp' ),
-      		'new_item_name'              => __( 'New Item Name', 'wpldp' ),
-      		'add_new_item'               => __( 'Add New Item', 'wpldp' ),
-      		'edit_item'                  => __( 'Edit Item', 'wpldp' ),
-      		'update_item'                => __( 'Update Item', 'wpldp' ),
-      		'view_item'                  => __( 'View Item', 'wpldp' ),
-      		'separate_items_with_commas' => __( 'Separate items with commas', 'wpldp' ),
-      		'add_or_remove_items'        => __( 'Add or remove items', 'wpldp' ),
-      		'choose_from_most_used'      => __( 'Choose from the most used', 'wpldp' ),
-      		'popular_items'              => __( 'Popular Items', 'wpldp' ),
-      		'search_items'               => __( 'Search Items', 'wpldp' ),
-      		'not_found'                  => __( 'Not Found', 'wpldp' ),
-      	);
-      	$rewrite = array(
-      		'slug'                       => '',
-      		'with_front'                 => false,
-      		'hierarchical'               => false,
-      	);
-      	$args = array(
-      		'labels'                     => $labels,
-      		'hierarchical'               => true,
-      		'public'                     => true,
-      		'show_ui'                    => true,
-      		'show_admin_column'          => true,
-      		'show_in_nav_menus'          => true,
-      		'show_tagcloud'              => true,
-      		// 'rewrite'                    => $rewrite,
-      	);
-      	register_taxonomy( 'ldp_container', 'ldp_resource', $args );
-
-      }
-
-      /**
-       	 * Adds a LDP Model field to our custom LDP containers taxonomy
-         * in creation mode
-         *
-       	 * @param int $term the concrete term
-       	 * @return void
-      	 */
-      function add_custom_tax_fields_oncreate($term) {
-        echo "<div class='form-field form-required term-model-wrap'>";
-        echo "<label for='ldp_model'>" . __('Model', 'wpldp'). "</label>";
-        echo "<textarea id='ldp_model' type='text' name='ldp_model' cols='40' rows='20'></textarea>";
-        echo "<p class='description'>" . __('The LDP-compatible JSON Model for this container', 'wpldp'). "</p>";
-        echo "</div>";
-      }
-
-      /**
-         * Adds a LDP Model field to our custom LDP containers taxonomy
-         * in edition mode
-         *
-       	 * @param int $term the concrete term
-       	 * @return void
-      	 */
-      function add_custom_tax_fields_onedit($term) {
-        $termId = $term->term_id;
-        $termMeta = get_option("ldp_container_$termId");
-        $ldpModel = stripslashes_deep($termMeta['ldp_model']);
-
-        echo "<tr class='form-field form-required term-model-wrap'>";
-        echo "<th scope='row'><label for='ldp_model_editor'>" . __('Model editor mode', 'wpldp'). "</label></th>";
-        echo "<td><div id='ldp_model_editor' style='width: 1000px; height: 400px;'></div></td>";
-        echo "<p class='description'>" . __('The LDP-compatible JSON Model for this container', 'wpldp'). "</p></td>";
-        echo "</tr>";
-        echo "<input type='hidden' id='ldp_model' name='ldp_model' value='$ldpModel'/>";
-
-        echo "</tr>";
-
-        echo '<script>
-                var container = document.getElementById("ldp_model_editor");
-                var options = {
-                  mode:"tree",
-                  modes: ["code", "form", "text", "tree", "view"],
-                  change: function () {
-                    var input = document.getElementById("ldp_model");
-                    if (input) {
-                      if (editor) {
-                        var json = editor.get();
-                        input.value = JSON.stringify(json);
-                      }
-                    }
-                  }
-                };
-                window.editor = new JSONEditor(container, options);
-
-                var json = ' . json_encode(json_decode($ldpModel)) . ';
-                editor.set(json);
-                editor.expandAll();
-              </script>';
-      }
-
-      /**
-       	 * Save the value of the posted custom field for the custom taxonomy
-         * in the options WP table
-       	 *
-       	 * @param type
-       	 * @return void
-      	 */
-      function save_custom_tax_field($termID) {
-        if (isset($_POST['ldp_model'])) {
-          $termMeta = get_option("ldp_container_$termID");
-          if (!is_array($termMeta)) {
-            $termMeta = array();
-          }
-
-          $termMeta['ldp_model'] = stripslashes_deep($_POST['ldp_model']);
-          update_option("ldp_container_$termID", $termMeta);
-        }
       }
 
       #############################
